# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Abrpardaz API",
        "description": "This is the official API which can be used by any user",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "Local server"
        },
        {
            "url": "http://abrpardaz.com/api/v1",
            "description": "The API endpoint"
        }
    ],
    "paths": {
...
        "/tickets/newTicket": {
            "post": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Crates a new ticket",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\TicketController::newTicket",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "machine",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/tickets/{id}/newReply": {
            "post": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Insert reply for a ticket",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\TicketController::newReply",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/tickets/{id}/close": {
            "put": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Closes a ticket",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\TicketController::close",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/tickets/{id}/show": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Show a ticket information",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\TicketController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login and copy the token into the below box",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}